implementing the addition functionality
adding 5 new procedures:
- InputNumber
- FormNumber
- ViewNumber
- View
- Exit

The user can select between 4 options (Add, Subtract, Multiply, Divide) using the numbers, if the user presses any other key or an invalid option the program will display a message and ask them to try again.
The user can add up to 2 numbers with a maximum result of a 5 digits  (99,999).
After the addition is complete, the result and a "goodbye" message will be displayed.
The user can press any key to exit the application.

The code has a lot of comments explaining each step

BUG: possible bug, if the user doesn't enter any number and press the "enter" key the program enters in an infinite loop